Tintin Colours

A list with 23 vectors (one per Tintin book) containing the five most frequent colors per cover. White tones were removed to balance the palettes. Alph-Art is Herge's unfinished book, it was published posthumously, so we don't consider it in the list.

tintin_colours

Format

A list with 23 vectors.

Source

Tintin covers obtained from Wikipedia and processed with ImageMagick.

Examples

head(tintin_colours)
#> $tintin_in_the_land_of_the_soviets
#> [1] "#165976" "#d04e66" "#ffd613" "#365158" "#3d809d"
#> 
#> $tintin_in_the_congo
#> [1] "#2ac4f6" "#b7da9a" "#ad653f" "#f2f311" "#947964"
#> 
#> $tintin_in_america
#> [1] "#87cceb" "#bbd87e" "#ec4f40" "#1c99b9" "#187352"
#> 
#> $cigars_of_the_pharaoh
#> [1] "#888a7f" "#7aa3d1" "#efae6a" "#b86d4d" "#e2a911"
#> 
#> $the_blue_lotus
#> [1] "#e30013" "#1a9abd" "#8d512d" "#de9c62" "#30515a"
#> 
#> $the_broken_ear
#> [1] "#8ba274" "#d8e492" "#89bfe5" "#5c8d78" "#b6c840"
#>