Tintin Head Trauma

Neurological traumas sustained by the subject (by book title)

tintin_head_trauma

Format

A data frame with 50 rows and 4 variables

Source

Cyr, Antoine, Louis-Olivier Cyr, and Claude Cyr. "Acquired growth hormone deficiency and hypogonadotropic hypogonadism in a subject with repeated head trauma, or Tintin goes to the neurologist." CMAJ 171, no. 12 (2004): 1433-1434. The book release year was obtained from Wikipedia.

Details

year Release year of the comic, first ocurrence was used for titles with a b&w and colour edition

book_title English title of the books

page Page number where the traumas happen

cause_of_injury What causes the trauma (i.e., explosion)

loss_of_consciousness_lengthLoss of consciousness length measured by the number of cartoon frames before subject returns to normal activity.

loss_of_consciousness_severityLoss of consciousness severity measured by the number of objects revolving above Tintin's head

Examples

tintin_head_trauma
#> # A tibble: 50 × 6
#>     year book_title           page cause_of_injury     loss_of_consciousness_l…¹
#>    <int> <fct>               <int> <fct>                                   <int>
#>  1  1930 Land of Soviets        21 Train accident                              1
#>  2  1930 Land of Soviets        86 Hit by a tree trunk                         2
#>  3  1930 Land of Soviets       105 Explosion                                   2
#>  4  1930 Land of Soviets       119 Blow from a club                            5
#>  5  1931 Tintin in the Congo    13 Struck by Snowy                             3
#>  6  1931 Tintin in the Congo    22 Fight with a lion                          11
#>  7  1931 Tintin in the Congo    32 Blow from a club                            6
#>  8  1931 Tintin in the Congo    43 Hit by a board                              6
#>  9  1932 Tintin in America       4 Car accident                                2
#> 10  1932 Tintin in America       5 Blow from a club                            3
#> # ℹ 40 more rows
#> # ℹ abbreviated name: ¹​loss_of_consciousness_length
#> # ℹ 1 more variable: loss_of_consciousness_severity <int>